// src/models/customization.model.ts
export interface CustomizationConfig {
    ConfigFile: string
    Username: string
    Password: string
    ProductType: string
    LayerType: string
    Key1: string
    Key2: string
    Key3: string
    CSGCode: string
    EpicorClientFolder: string
    Company: string
    Folder: string
    Action: string
    ProjectFolder: string
    DLLLocation: string
    DNSpy: any
    Version: number
    Encrypted: string
    Tracing: string
    NewConfig: string
    Temp: string
  }
  